Refactor file ChecksumMD5
