+++ src/main/java/org/fundacionjala/converter/controller/UploadFileController.java
@@ -10,7 +10,7 @@ package org.fundacionjala.converter.controller;
 
 import org.fundacionjala.converter.controller.exceptions.NonExistentException;
 import org.fundacionjala.converter.database.exception.NullAttributeException;
-import org.fundacionjala.converter.model.ChecksumMD5;
+import org.fundacionjala.converter.model.utility.ChecksumMD5;
 import org.fundacionjala.converter.database.entity.File;
 import org.fundacionjala.converter.controller.service.FileService;
 import org.fundacionjala.converter.controller.service.FileUploadService;
